// src/types/commonTypes.ts
export type ResponseData = {
  avg_price: string;
  cnc_used_quantity: number;
  collateral_qty: number;
  collateral_type: string;
  collateral_update_qty: number;
  company_name: string;
  haircut: number;
  holdings_update_qty: number;
  isin: string;
  product: string;
  quantity: number;
  symbol: string;
  t1_holding_qty: number;
  token_bse: string;
  token_nse: string;
  withheld_collateral_qty: number;
  withheld_holding_qty: number;
  ltp: number;
  close: number;
};